+++ test/filters/handler/timeout.js
@@ -5,15 +5,16 @@ var util = require('util');
 var mockSession = {
   key : "123"
 };
+let msg = {__route__: 'mockTest'};
 
-var WAIT_TIME = 100;
-describe("#serialFilter",function(){
+var WAIT_TIME = 1500;
+describe("#timeoutFilter",function(){
   it("should do before filter ok",function(done){
     var service = new FilterService();
-    var filter = timeoutFilter();
+    var filter = timeoutFilter(1000);
     service.before(filter);
 
-    service.beforeFilter(null,mockSession,function(){
+    service.beforeFilter(msg,mockSession,function(){
       should.exist(mockSession);
 
       should.exist(mockSession.__timeout__);
@@ -23,24 +24,22 @@ describe("#serialFilter",function(){
 
   it("should do after filter by doing before filter ok",function(done){
     var service = new FilterService();
-    var filter = timeoutFilter();
+    var filter = timeoutFilter(1000);
     var _session ;
     service.before(filter);
+    service.after(filter);
+    
 
-    service.beforeFilter(null,mockSession,function(){
+    service.beforeFilter(msg,mockSession,function(){
       should.exist(mockSession);
       should.exist(mockSession.__timeout__);
       _session = mockSession;
+
+      service.afterFilter(null,msg,mockSession,null,function(){
+        should.exist(mockSession);
+        should.strictEqual(mockSession,_session);
+        done();
+      });
     });
-
-    service.after(filter);
-
-    service.afterFilter(null,null,mockSession,null,function(){
-      should.exist(mockSession);
-      should.strictEqual(mockSession,_session);
-    });
-
-    setTimeout(done,WAIT_TIME);
-    done();
   });
 });