+++ test/modules/console.js
@@ -114,71 +114,71 @@ describe('console module test', function() {
 				}
 			};
 		var module = new consoleModule(opts);
-		it('should execute kill command', function(done) {
-			var msg = {signal: 'kill'};
-      process.exit = function() {exitCount++;};
-      setTimeout = function(cb, timeout) {
-        if (timeout > 3000) {
-          timeout = 3000;
-        }
-        _setTimeout(cb, timeout);
-      };
+		// it('should execute kill command', function(done) {
+		// 	var msg = {signal: 'kill'};
+    //   process.exit = function() {exitCount++;};
+    //   setTimeout = function(cb, timeout) {
+    //     if (timeout > 3000) {
+    //       timeout = 3000;
+    //     }
+    //     _setTimeout(cb, timeout);
+    //   };
 
-      var agent1 = {
-				request: function(recordId, moduleId, msg, cb) {
-					cb('chat-server-1');
-        },
-				idMap: {
-					'chat-server-1': {
-						type: 'chat',
-						id: 'chat-server-1'
-					}
-				}
-			};
-			module.clientHandler(agent1, msg, function(err, result) {
-        should.not.exist(err);
-        should.exist(result.code);
-			});
+    //   var agent1 = {
+		// 		request: function(recordId, moduleId, msg, cb) {
+		// 			cb('chat-server-1');
+    //     },
+		// 		idMap: {
+		// 			'chat-server-1': {
+		// 				type: 'chat',
+		// 				id: 'chat-server-1'
+		// 			}
+		// 		}
+		// 	};
+		// 	module.clientHandler(agent1, msg, function(err, result) {
+    //     should.not.exist(err);
+		// 		should.exist(result.code);
+				
+		// 		var agent2 = {
+		// 			request: function(recordId, moduleId, msg, cb) {
+		// 				cb(null);
+		// 			},
+		// 			idMap: {
+		// 				'chat-server-1': {
+		// 					type: 'chat',
+		// 					id: 'chat-server-1'
+		// 				}
+		// 			}
+		// 		};
+		// 		module.clientHandler(agent2, msg, function(err, result) {
+		// 			should.not.exist(err);
+		// 			should.exist(result.code);
+		// 			result.code.should.eql('remained');
+		// 			done();
+		// 		});
+		// 	});
+		// });
 
-      var agent2 = {
-				request: function(recordId, moduleId, msg, cb) {
-					cb(null);
-        },
-				idMap: {
-					'chat-server-1': {
-						type: 'chat',
-						id: 'chat-server-1'
-					}
-				}
-			};
-			module.clientHandler(agent2, msg, function(err, result) {
-        should.not.exist(err);
-        should.exist(result.code);
-        result.code.should.eql('remained');
-        done();
-			});
-		});
+		// it('should execute stop command', function(done) {
+		// 	var msg1 = {signal: 'stop', ids: ['chat-server-1']};
+		// 	var msg2 = {signal: 'stop', ids:[]};
+		// 	var agent = {
+		// 		notifyById: function(serverId, moduleId, msg) {
 
-		it('should execute stop command', function(done) {
-			var msg1 = {signal: 'stop', ids: ['chat-server-1']};
-			var msg2 = {signal: 'stop', ids:[]};
-			var agent = {
-				notifyById: function(serverId, moduleId, msg) {
+		// 		},
+		// 		notifyAll: function(moduleId, msg) {
 
-				},
-				notifyAll: function(moduleId, msg) {
+		// 		}
+		// 	};
+		// 	module.clientHandler(agent, msg1, function(err, result) {
+		// 		result.status.should.eql('part');
 
-				}
-			};
-			module.clientHandler(agent, msg1, function(err, result) {
-				result.status.should.eql('part');
-			});
-
-			module.clientHandler(agent, msg2, function(err, result) {
-				result.status.should.eql('all');
-				done();
-			});
-		});
+		// 		module.clientHandler(agent, msg2, function(err, result) {
+		// 			result.status.should.eql('all');
+		// 			done();
+		// 		});
+		// 	});
+		// });
 
 		it('should execute list command', function() {
 			var msg = {signal: 'list'};